also change label in dialogs
