+++ src/renderer/ANumbersCellRenderer.ts
@@ -1,7 +1,7 @@
 import {IDataRow, INumbersColumn, EAdvancedSortMethod, IOrderedGroup} from '../model';
 import {IRenderContext, IImposer} from './interfaces';
 import {renderMissingCanvas, renderMissingDOM} from './missing';
-import {ISequence, boxplotBuilder} from '../internal';
+import {ISequence, boxplotBuilder, getSortLabel} from '../internal';
 
 /** @internal */
 export abstract class ANumbersCellRenderer {
@@ -90,29 +90,5 @@ export function matchRows(n: HTMLElement | SVGElement, length: number, template:
     children.slice(length).forEach((c) => c.remove());
   } else if (length > children.length) {
     n.insertAdjacentHTML('beforeend', template.repeat(length - children.length));
-  }
-}
-
-/**
- * generates a label for the given sort method
- * @internal
- * @param method sort method
- */
-export function getSortLabel(method: string) {
-  switch (method) {
-    case 'min':
-      return 'Mininum';
-    case 'max':
-      return 'Maximum';
-    case 'median':
-      return 'Median';
-    case 'mean':
-      return 'Mean';
-    case 'q1':
-      return '25% Quantile';
-    case 'q3':
-      return '75% Quantile';
-    default:
-      return String(method);
   }
 }
